# product-api/swagger.yaml
...
info: {}
paths:
...
    /products/{id}:
        delete:
            description: Return a list of products
            operationId: deleteProduct
            parameters:
                - format: int64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
            responses:
                "201":
                    $ref: '#/responses/noContent'
            tags:
                - products
responses:
    noContent:
        description: ""
...
swagger: "2.0"